2nd round patches for openGauss 1.1.0

+++ src/gausskernel/storage/mot/jit_exec/src/jit_llvm_blocks.cpp
@@ -580,7 +580,7 @@ static llvm::Value* ProcessOpExpr(
     JitLlvmCodeGenContext* ctx, const OpExpr* op_expr, int& result_type, int arg_pos, int depth, int* max_arg)
 {
     llvm::Value* result = nullptr;
-    MOT_LOG_DEBUG("%*s --> Processing OP %d expression", depth, "", (int)op_expr->opfuncid);
+    MOT_LOG_DEBUG("%*s --> Processing OP %u expression", depth, "", op_expr->opfuncid);
     if (depth > MOT_JIT_MAX_EXPR_DEPTH) {
         MOT_LOG_TRACE("Cannot process expression: Expression exceeds depth limit %d", (int)MOT_JIT_MAX_EXPR_DEPTH);
         return nullptr;
@@ -615,11 +615,11 @@ static llvm::Value* ProcessOpExpr(
         APPLY_OPERATORS()
 
         default:
-            MOT_LOG_TRACE("Unsupported operator function type: %d", op_expr->opfuncid);
+            MOT_LOG_TRACE("Unsupported operator function type: %u", op_expr->opfuncid);
             break;
     }
 
-    MOT_LOG_DEBUG("%*s <-- Processing OP %d expression result: %p", depth, "", (int)op_expr->opfuncid, result);
+    MOT_LOG_DEBUG("%*s <-- Processing OP %u expression result: %p", depth, "", op_expr->opfuncid, result);
     return result;
 }
